requirements.txt
Django==2.2.16
django-ckeditor==6.0.0
django-js-asset==1.2.2
pytz==2020.1
sqlparse==0.4.1
